Add files via upload
One of the most difficult topics in the field of intelligent transportation systems is
the automatic interpretation of driver behavior. This research examines the recognition of
distracted driver posture as part of the recognition of human action. Many car accidents
have been reported, caused by distracted drivers. Our objective was to improve the
performance of detection of distracted driver actions.
In our study, we tackle the problem of distracted driving by aiming to build a robust
multi-class classifier to detect and identify different forms. driver inattention using the
State Farm distracted driving dataset. We use a model built from zero, pre-trained
image classification models like VGG16 and ResNet50, and image preprocessing based
on OpenCV. The results give an accuracy of 98% with the model built from zero which
outperformed the pre-trained models which did not exceed a precision of 63% with ResNet
and 43,5% with VGG-16.